Fix some warning

+++ src/views/AppView.vala
@@ -68,13 +68,6 @@ public class Gitignore.Views.AppView : Gtk.Grid {
             copy_toast.send_notification ();
         });
 
-        copy_button.clicked.connect (() => {
-            Gtk.Clipboard clipboard = Gtk.Clipboard.get_for_display (display, Gdk.SELECTION_CLIPBOARD);
-            clipboard.set_text (gitignore_view.source_buffer.text, -1);
-            copy_toast.valign = Gtk.Align.END;
-            copy_toast.send_notification ();
-        });
-
         save_button.clicked.connect (() => {
             var filech = Utils.new_file_chooser_dialog ( _("Save File"), null);
             filech.do_overwrite_confirmation = true;
@@ -83,7 +76,7 @@ public class Gitignore.Views.AppView : Gtk.Grid {
             if (filech.run () == Gtk.ResponseType.ACCEPT) {
                 try {
                     var data_file = File.new_for_uri (filech.get_current_folder_uri () +"/.gitignore");
-                    
+
                     var file_stream = data_file.create (FileCreateFlags.NONE);
 
                     if (data_file.query_exists ()) {
@@ -92,7 +85,7 @@ public class Gitignore.Views.AppView : Gtk.Grid {
 
                     var data_stream = new DataOutputStream (file_stream);
                     data_stream.put_string (gitignore_view.source_buffer.text);
-                    
+
                     file_created_toast.valign = Gtk.Align.END;
                     file_created_toast.send_notification ();
                 } catch (Error e) {
@@ -103,7 +96,7 @@ public class Gitignore.Views.AppView : Gtk.Grid {
             filech.destroy ();
         });
 
-        toggle_buttons ();
+        update_tags ();
     }
 
     construct {
@@ -116,8 +109,6 @@ public class Gitignore.Views.AppView : Gtk.Grid {
         content_box.valign = Gtk.Align.CENTER;
 
         tag_grid = new Gtk.Grid ();
-
-        update_tags ();
 
         save_button = new Gtk.Button.from_icon_name ("document-save-as", Gtk.IconSize.LARGE_TOOLBAR);
         save_button.set_tooltip_text (_("Save as file"));
@@ -132,7 +123,7 @@ public class Gitignore.Views.AppView : Gtk.Grid {
         generate_gitignore_button.get_style_context ().add_class (Gtk.STYLE_CLASS_SUGGESTED_ACTION);
         generate_gitignore_button.get_style_context ().add_class ("flat");
 
-        content_box.pack_start (tag_grid, false, false, 0);
+       content_box.pack_start (tag_grid, false, false, 0);
         content_box.pack_end (generate_gitignore_button, false, false, 0);
         content_box.pack_end (bookmark_button, false, false, 0);
         content_box.pack_end (copy_button, false, false, 0);
@@ -145,8 +136,8 @@ public class Gitignore.Views.AppView : Gtk.Grid {
         stack.transition_type = Gtk.StackTransitionType.SLIDE_LEFT_RIGHT;
         stack.add_titled ( welcome_view, "welcome_view_stack", _("Welcome View"));
         stack.add_titled ( gitignore_view, "gitignore_view_stack", _("Gitignore View"));
-        stack.show_all ();
         stack.visible_child_name = "welcome_view_stack";
+        stack.show_all ();
 
         copy_toast = new Granite.Widgets.Toast (_("Copied content to clipboard"));
         file_created_toast = new Granite.Widgets.Toast (_("File successfully created."));
@@ -163,8 +154,10 @@ public class Gitignore.Views.AppView : Gtk.Grid {
 
         string[] data = settings.get_strv ("selected-langs");
 
-        for (int i = 0; i < data.length + 1; i++) {
-            tag_grid.remove_column (i);
+        if (data != null && data.length > 0) {
+            for (int i = 0; i < data.length + 1; i++) {
+                tag_grid.remove_column (i);
+            }
         }
 
         var children = tag_grid.get_children ();
@@ -172,12 +165,14 @@ public class Gitignore.Views.AppView : Gtk.Grid {
             tag_grid.remove (element);
         }
 
-        for (int i = 0; i < data.length; i++) {
-            var tag = new Gitignore.Widgets.Tag (data[i]);
-            tag_grid.attach (tag, i, 0);
-            tag.tag_deleted.connect (() => {
-                update_tags ();
-            });
+        if (data != null && data.length > 0) {
+            for (int i = 0; i < data.length; i++) {
+                var tag = new Gitignore.Widgets.Tag (data[i]);
+                tag_grid.attach (tag, i, 0);
+                tag.tag_deleted.connect (() => {
+                    update_tags ();
+                });
+            }
         }
 
         tag_grid.show_all ();
@@ -192,27 +187,27 @@ public class Gitignore.Views.AppView : Gtk.Grid {
 
         if (stack.visible_child_name == "welcome_view_stack") {
             save_button.set_sensitive (false);
-            save_button.set_opacity (0);
             copy_button.set_sensitive (false);
+            save_button.set_opacity (0);
             copy_button.set_opacity (0);
 
-            if (data.length > 0) {
-                generate_gitignore_button.set_sensitive (true);
-            } else {
-                generate_gitignore_button.set_sensitive (false);
-            }
-        } else if (stack.visible_child_name == "gitignore_view_stack") {
-            if (data.length > 0) {
+            if (data != null && data.length > 0) {
                 generate_gitignore_button.set_sensitive (true);
             } else {
                 generate_gitignore_button.set_sensitive (false);
             }
 
+        } else if (stack.visible_child_name == "gitignore_view_stack") {
             copy_button.set_sensitive (true);
             save_button.set_sensitive (true);
             save_button.set_opacity (1);
             copy_button.set_opacity (1);
-            generate_gitignore_button.set_opacity (1);
+
+            if (data != null && data.length > 0) {
+                generate_gitignore_button.set_sensitive (true);
+            } else {
+                generate_gitignore_button.set_sensitive (false);
+            }
         }
     }
 }
